+++ frontend/wailsjs/go/services/SettingsService.js
@@ -50,6 +50,6 @@ export function Startup(arg1) {
   return window['go']['services']['SettingsService']['Startup'](arg1);
 }
 
-export function TestConnection(arg1, arg2) {
-  return window['go']['services']['SettingsService']['TestConnection'](arg1, arg2);
+export function TestConnection(arg1, arg2, arg3) {
+  return window['go']['services']['SettingsService']['TestConnection'](arg1, arg2, arg3);
 }